code/back-end/requirement.txt
nltk
tweepy
MeaningCloud-python
pandas
django
djangorestframework